Moved timeline model out of components
